insert user, insert milestone wip

+++ dto/Goal.go
@@ -7,6 +7,6 @@ import (
 type Goal struct {
 	ID        int            `db:"id"`
 	Challenge int            `db:"challenge"`
-	Typ       int            `db:"typ"`
+	Typ       byte           `db:"typ"`
 	Descript  sql.NullString `db:"descript"`
 }
